Update CodeFund embed script (#1492)
<!-- Is it a Bug fix, feature, docs update, ... -->
This is a feature and doc update.

<!-- You can also link to an open issue here -->
The current CodeFund embed script is using the legacy (deprecated) API and URL, which has not been whitelisted with AcceptableAds.com.

<!-- if this is a feature change -->
**What is the new behavior?**

This PR updates the API endpoint and URL to enable whitelisting and in-app configuration.

I also updated the links on the README and contributors to use the new URL: codefund.app.

<!-- Have you done all of these things?  -->
**Checklist**:
<!-- add "N/A" to the end of each line that's irrelevant to your changes -->
<!-- to check an item, place an "x" in the box like so: "- [x] Documentation" -->
- [x] Documentation
- [ ] Tests (N/A)
- [ ] Ready to be merged <!-- In your opinion, is this ready to be merged as soon as it's reviewed? -->
- [ ] Added myself to contributors table <!-- this is optional, see the contributing guidelines for instructions -->

<!-- feel free to add additional comments -->

I was unable to test this locally. Please test the new implementation prior to merge. This can be done by going to a sandbox unauthenticated and viewing the ad in the bottom left.

<!-- Thank you for contributing! -->