New Problem View (#1794)
With this update I move the corrections generated by the linter from the state of our editor to the global state (in case of `app` in Cerebral, in `embed` we use a variable). This way we can show the corrections of the linter in the `Problems` view, which got a small redesign as well to be more useful. In the future we might decide to also include the `corrections`/`errors` generated by VSCode, or we might move the source of truth of this state to VSCode entirely.

- [x] Moves lint errors to global state
- [x] Removes the need of `moduleId` for corrections and  errors
- [x] Removes glyphs (we didn't use it)
- [x] Adds ability to open specific line of a file

![image](https://user-images.githubusercontent.com/587016/56470576-18eb2f80-6448-11e9-95ae-4cdac0d60c8f.png)