New Privacy Policy (#3815)
Major changes to our currently outdated and privacy statement containing inaccurate statements. Main reason for the rewrites are:
* Add European or EEA focus and compliance with laws and regulations;
* Support introduction of upcoming email campaigns;
* Tell you who we partner with;
* Provide more accurate information on how we treat personal data.

This PR also introduces homepage and editor notifications, informing viewers about the applicability of the policy.

Privacy is an ongoing concern for CodeSandbox and our next update is likely to contain more information on international relations, such as there is the Privacy Shield Framework between US-EU. Please do reach out to us with questions.