Fix tag fetching (#5924)
We used to use jsdelivr for resolving a tag to an absolute version, but
it turns out that jsdelivr caches the result. This is unintended
behavior, because when a tag updates people will see different results
based on their browser cache. This PR makes us default to unpkg instead,
which has a better caching mechanism.